feat: - produced eggs go straight to the PC - badges now speed up egg hatching and laying fix: - bike type swap and bike item text was borked

+++ include/egg_hatch.h
@@ -4,7 +4,7 @@
 void ScriptHatchMon(void);
 bool8 CheckDaycareMonReceivedMail(void);
 void EggHatch(void);
-u8 GetEggCyclesToSubtract(void);
+u32 GetEggCyclesToSubtract(void);
 u16 CountPartyAliveNonEggMons(void);
 
 #endif // GUARD_EGG_HATCH_H
